refactor(network-interception): add package exports for type subpaths
Declare ./types and ./types/external export entries so consumers import
@packages/network-interception/types instead of deep lib/ paths. Add
typesVersions for TypeScript resolution under classic moduleResolution.
Update CLI sync-typedefs to resolve the external types export.
