+++ cli/scripts/sync-typedefs.ts
@@ -6,7 +6,10 @@ import fs from 'fs-extra'
 import { includeTypes } from './utils'
 import shell from 'shelljs'
 import { join } from 'path'
+import { createRequire } from 'module'
 import resolvePkg from 'resolve-pkg'
+
+const require = createRequire(import.meta.url)
 
 shell.set('-v') // verbose
 shell.set('-e') // any error is fatal
@@ -73,7 +76,7 @@ shell.sed('-i', 'from "sinon";', 'from "../sinon";', sinonChaiFilename)
 
 // copy experimental network stubbing type definitions
 // so users can import: `import 'cypress/types/net-stubbing'`
-fs.copySync(resolvePkg('@packages/network-interception/lib/types/external-types.ts'), 'types/net-stubbing.d.ts')
+fs.copySync(require.resolve('@packages/network-interception/types/external'), 'types/net-stubbing.d.ts')
 
 // https://github.com/cypress-io/cypress/issues/18069
 // To avoid type clashes, some files should be commented out entirely by patch-package
