+++ packages/driver/cypress/e2e/commands/net_stubbing.cy.ts
@@ -1,5 +1,5 @@
 import { getDisplayUrlMatcher } from '../../../src/cy/net-stubbing/route-matcher-log'
-import type { RouteMatcherOptions } from '@packages/network-interception/lib/types/external-types'
+import type { RouteMatcherOptions } from '@packages/network-interception/types/external'
 
 const testFail = (cb, expectedDocsUrl = 'https://on.cypress.io/intercept') => {
   cy.on('fail', (err) => {
