+++ packages/driver/src/cy/net-stubbing/events/after-response.ts
@@ -1,4 +1,4 @@
-import type { CyHttpMessages } from '@packages/network-interception/lib/types'
+import type { CyHttpMessages } from '@packages/network-interception/types'
 import type { HandlerFn } from '.'
 import { parseJsonBody } from './utils'
