+++ packages/driver/src/cy/net-stubbing/events/response.ts
@@ -3,7 +3,7 @@ import _ from 'lodash'
 import {
   CyHttpMessages,
   SERIALIZABLE_RES_PROPS,
-} from '@packages/network-interception/lib/types'
+} from '@packages/network-interception/types'
 import {
   validateStaticResponse,
   parseStaticResponseShorthand,
