+++ packages/driver/src/cypress/proxy-logging.ts
@@ -1,5 +1,5 @@
 import _ from 'lodash'
-import type { Interception, Route } from '@packages/network-interception/lib/types'
+import type { Interception, Route } from '@packages/network-interception/types'
 import type { BrowserPreRequest, BrowserResponseReceived, RequestError } from '@packages/proxy/lib/types'
 import $errUtils from './error_utils'
 import Debug from 'debug'
