+++ packages/net-stubbing/lib/server/handle-intercept-response.ts
@@ -7,7 +7,7 @@ import type { ResponseMiddleware } from '@packages/proxy'
 import {
   CyHttpMessages,
   SERIALIZABLE_RES_PROPS,
-} from '@packages/network-interception/lib/types'
+} from '@packages/network-interception/types'
 import {
   getBodyStream,
   mergeDeletedHeaders,
